issues/41 fix rabbitmq routing & schedule deletion error

+++ src/Focus.Service.ReportScheduler/Application/Events/DeleteOutdatedSchedules.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Focus.Application.Common.Messages.Events;
 using Focus.Service.ReportScheduler.Application.Services;
 using MediatR;
+using Microsoft.Extensions.Logging;
 
 namespace Focus.Service.ReportScheduler.Application.Events
 {
@@ -12,27 +13,35 @@ namespace Focus.Service.ReportScheduler.Application.Events
     public class DeleteOutdatedSchedules : INotificationHandler<NewDay>
     {
         private readonly IReportScheduleRepository _repository;
+        private readonly ILogger<DeleteOutdatedSchedules> _logger;
 
-        public DeleteOutdatedSchedules(IReportScheduleRepository repository)
+        public DeleteOutdatedSchedules(IReportScheduleRepository repository, ILogger<DeleteOutdatedSchedules> logger)
         {
             _repository = repository;
+            _logger = logger;
         }
 
         public async Task Handle(NewDay notification, CancellationToken cancellationToken)
         {
+            _logger.LogInformation("Initialized deletion of outdated schedules");
+
             try
             {
                 var schedules = await _repository.GetReportSchedulesAsync();
 
                 var outdated = schedules
-                    .Where(s => s.EmissionEnd.Date > DateTime.Now.Date.AddDays(15))
+                    .Where(s => s.EmissionEnd.Date < DateTime.Now.Date.AddDays(15))
                     .Select(s => s.Id);
 
-                await _repository.DeleteSchedulesAsync(outdated);
+                _logger.LogInformation($"Schedules to delete: {outdated.Count()}");
+
+                if (outdated != null && outdated.Count() > 0)
+                    await _repository.DeleteSchedulesAsync(outdated);
             }
-            catch (Exception)
+            catch (Exception e)
             {
-                throw;
+                _logger.LogError(e, e.Message);
+                throw e;
             }
         }
     }
